requirements.txt
requests==2.28.2
dnspython==2.3.0
